fix(derive): data availability abstractions
