+++ crates/derive-alloy/src/pipeline.rs
@@ -84,8 +84,11 @@ mod tests {
         let beacon_client = OnlineBeaconClient::new_http("http://127.0.0.1:5555".into());
         let blob_provider = OnlineBlobProvider::new(beacon_client, None, None);
         let blob_provider = OnlineBlobProviderWithFallback::new(blob_provider, None);
-        let dap_source =
-            EthereumDataSource::new(chain_provider.clone(), blob_provider, &rollup_config);
+        let dap_source = EthereumDataSource::new_from_parts(
+            chain_provider.clone(),
+            blob_provider,
+            &rollup_config,
+        );
         let builder = StatefulAttributesBuilder::new(
             rollup_config.clone(),
             l2_chain_provider.clone(),
