+++ crates/derive/src/stages/l1_traversal.rs
@@ -25,9 +25,9 @@ pub struct L1Traversal<Provider: ChainProvider> {
     /// The current block in the traversal stage.
     pub block: Option<BlockInfo>,
     /// The data source for the traversal stage.
-    data_source: Provider,
+    pub data_source: Provider,
     /// Signals whether or not the traversal stage is complete.
-    done: bool,
+    pub done: bool,
     /// The system config.
     pub system_config: SystemConfig,
     /// A reference to the rollup config.
