+++ crates/derive/src/traits/mod.rs
@@ -11,7 +11,7 @@ mod attributes;
 pub use attributes::{AttributesBuilder, AttributesProvider, NextAttributes};
 
 mod data_sources;
-pub use data_sources::{AsyncIterator, BlobProvider, DataAvailabilityProvider};
+pub use data_sources::{BlobProvider, DataAvailabilityProvider};
 
 mod reset;
 pub use reset::ResetProvider;
